refactor(composer): toggle AI draft mode without direct submit and extract subroutines
